+++ src/gcovr/formats/__init__.py
@@ -100,6 +100,11 @@ def read_reports(options: Options) -> CoverageContainer:
     else:
         covdata = GcovHandler(options).read_report()
 
+    if not covdata:
+        LOGGER.warning(
+            "All coverage data is filtered out. Please check your paths and filters."
+        )
+
     if options.include_search_filter:
         for search_path in options.search_paths or [options.root]:
             LOGGER.debug("Search for included files in %s", search_path)